Restructuring packages in app to be a little more organized.
