+++ app/src/main/java/com/kszalai/billtracker/bills/common/models/CreditCardDetails.kt
@@ -1,4 +1,4 @@
-package com.kszalai.billtracker.models
+package com.kszalai.billtracker.bills.common.models
 
 data class CreditCardDetails(val limit: Double,
                              val apr: Double)
