+++ app/src/main/java/com/kszalai/billtracker/bills/common/models/CreditCardLimit.kt
@@ -1,4 +1,4 @@
-package com.kszalai.billtracker.models
+package com.kszalai.billtracker.bills.common.models
 
 data class CreditCardLimit(val creditLimit: Double,
                            val apr: Double)
