+++ app/src/main/java/com/kszalai/billtracker/bills/details/BillDetailViewModel.kt
@@ -3,7 +3,7 @@ package com.kszalai.billtracker.bills.details
 import androidx.compose.runtime.mutableStateOf
 import androidx.lifecycle.ViewModel
 import com.kszalai.billtracker.bills.common.repo.BillRepo
-import com.kszalai.billtracker.models.BillObject
+import com.kszalai.billtracker.bills.common.models.BillObject
 import dagger.hilt.android.lifecycle.HiltViewModel
 import javax.inject.Inject
