+++ app/src/main/java/com/kszalai/billtracker/bills/details/BillDetailsUIState.kt
@@ -1,6 +1,6 @@
 package com.kszalai.billtracker.bills.details
 
-import com.kszalai.billtracker.models.BillObject
+import com.kszalai.billtracker.bills.common.models.BillObject
 
 data class BillDetailsUIState(
     val loading: Boolean = false,
