--- app/src/main/java/com/kszalai/billtracker/models/BillDetailRecyclerViewItems.kt
@@ -1,5 +1,0 @@
-package com.kszalai.billtracker.models
-
-enum class BillDetailRecyclerViewItems {
-    PaymentHeader
-}
